split into retrieval, upsert repo
